Fix problems with tx<->arb functions
These functions weren't handling conversion on arbs that contained a mixture
of other arbs and :content/text entitites. I think I wasn't clear on whether
it was possible to have just :content/text, not wrapped in an arb. The fixes
here assume that it is; that you can have either an arb or :content/text in
the :arb/value array. It's possible this isn't right, and I may need to change
later.